Moving two functions from RSC to RestRequest
Moving setDigestAlgorithm() and getDigest() from ReadableStreamChannel to RestRequest
After examining the use cases, this seesm like the right move since not all channels
need to support digests. It is an overhead for many implementations of ReadableStreamChannel.
