+++ ambry-api/src/test/java/com.github.ambry/router/ByteBufferRSC.java
@@ -16,8 +16,6 @@ package com.github.ambry.router;
 import java.io.IOException;
 import java.nio.ByteBuffer;
 import java.nio.channels.ClosedChannelException;
-import java.security.MessageDigest;
-import java.security.NoSuchAlgorithmException;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.concurrent.Future;
@@ -63,10 +61,6 @@ public class ByteBufferRSC implements ReadableStreamChannel {
 
   protected final ByteBuffer buffer;
   protected final int size;
-  private final int startPos;
-
-  private MessageDigest digest = null;
-  private byte[] digestBytes = null;
 
   /**
    * Constructs a {@link ReadableStreamChannel} whose read operations return data from the provided {@code buffer}.
@@ -75,7 +69,6 @@ public class ByteBufferRSC implements ReadableStreamChannel {
   public ByteBufferRSC(ByteBuffer buffer) {
     this.buffer = buffer;
     size = buffer.remaining();
-    startPos = buffer.position();
   }
 
   @Override
@@ -102,34 +95,6 @@ public class ByteBufferRSC implements ReadableStreamChannel {
     }
     onEventComplete(Event.ReadInto);
     return future;
-  }
-
-  @Override
-  public void setDigestAlgorithm(String digestAlgorithm)
-      throws NoSuchAlgorithmException {
-    if (digest != null && digest.getAlgorithm().equals(digestAlgorithm)) {
-      onEventComplete(Event.SetDigestAlgorithm);
-      return;
-    }
-    digest = MessageDigest.getInstance(digestAlgorithm);
-    digestBytes = null;
-    onEventComplete(Event.SetDigestAlgorithm);
-  }
-
-  @Override
-  public byte[] getDigest() {
-    if (digest == null) {
-      onEventComplete(Event.GetDigest);
-      return null;
-    } else if (digestBytes == null) {
-      int savedPosition = buffer.position();
-      buffer.position(startPos);
-      digest.update(buffer);
-      buffer.position(savedPosition);
-      digestBytes = digest.digest();
-    }
-    onEventComplete(Event.GetDigest);
-    return digestBytes;
   }
 
   @Override
