+++ ambry-api/src/test/java/com.github.ambry/router/ByteRangeTest.java
@@ -16,7 +16,9 @@ package com.github.ambry.router;
 
 import org.junit.Test;
 
-import static org.junit.Assert.*;
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertFalse;
+import static org.junit.Assert.fail;
 
 
 /**
