+++ ambry-api/src/test/java/com.github.ambry/router/GetBlobOptionsTest.java
@@ -16,7 +16,8 @@ package com.github.ambry.router;
 
 import org.junit.Test;
 
-import static org.junit.Assert.*;
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertFalse;
 
 
 /**
