+++ ambry-commons/src/main/java/com.github.ambry.commons/ByteBufferReadableStreamChannel.java
@@ -20,8 +20,6 @@ import com.github.ambry.router.ReadableStreamChannel;
 import java.io.IOException;
 import java.nio.ByteBuffer;
 import java.nio.channels.ClosedChannelException;
-import java.security.MessageDigest;
-import java.security.NoSuchAlgorithmException;
 import java.util.concurrent.Future;
 import java.util.concurrent.atomic.AtomicBoolean;
 
@@ -34,10 +32,6 @@ public class ByteBufferReadableStreamChannel implements ReadableStreamChannel {
   private final AtomicBoolean channelEmptied = new AtomicBoolean(false);
   private final ByteBuffer buffer;
   private final int size;
-  private final int startPos;
-
-  private MessageDigest digest;
-  private byte[] digestBytes;
 
   /**
    * Constructs a {@link ReadableStreamChannel} whose read operations return data from the provided {@code buffer}.
@@ -46,7 +40,6 @@ public class ByteBufferReadableStreamChannel implements ReadableStreamChannel {
   public ByteBufferReadableStreamChannel(ByteBuffer buffer) {
     this.buffer = buffer;
     size = buffer.remaining();
-    startPos = buffer.position();
   }
 
   @Override
@@ -71,48 +64,6 @@ public class ByteBufferReadableStreamChannel implements ReadableStreamChannel {
       future = asyncWritableChannel.write(buffer, callback);
     }
     return future;
-  }
-
-  /**
-   * {@inheritDoc}
-   * <p/>
-   * This implementation supports setting the digest algorithm at any point in the lifecycle of the object. Digest
-   * algorithms can be changed at will and a subsequent call to {@link #getDigest()} will get the digest as computed
-   * by the algorithm that was set most recently.
-   * @param digestAlgorithm the digest algorithm to use.
-   * @throws NoSuchAlgorithmException if the {@code digestAlgorithm} does not exist or is not supported.
-   */
-  @Override
-  public void setDigestAlgorithm(String digestAlgorithm)
-      throws NoSuchAlgorithmException {
-    if (digest != null && digest.getAlgorithm().equals(digestAlgorithm)) {
-      return;
-    }
-    digest = MessageDigest.getInstance(digestAlgorithm);
-    digestBytes = null;
-  }
-
-  /**
-   * {@inheritDoc}
-   * <p/>
-   * This implementation supports getting the digest at any point in the lifecycle of the object. The digest is always
-   * that of all the data in the channel regardless of how much data has been consumed by
-   * {@link #readInto(AsyncWritableChannel, Callback)}.
-   * @return the digest as computed by the digest algorithm set through {@link #setDigestAlgorithm(String)}. If none
-   * was set, {@code null}.
-   */
-  @Override
-  public byte[] getDigest() {
-    if (digest == null) {
-      return null;
-    } else if (digestBytes == null) {
-      int savedPosition = buffer.position();
-      buffer.position(startPos);
-      digest.update(buffer);
-      buffer.position(savedPosition);
-      digestBytes = digest.digest();
-    }
-    return digestBytes;
   }
 
   @Override
