+++ ambry-commons/src/test/java/com.github.ambry.commons/ReadableStreamChannelInputStreamTest.java
@@ -218,16 +218,6 @@ class IncompleteReadReadableStreamChannel implements ReadableStreamChannel {
   }
 
   @Override
-  public void setDigestAlgorithm(String digestAlgorithm) {
-    throw new IllegalStateException("Not implemented");
-  }
-
-  @Override
-  public byte[] getDigest() {
-    throw new IllegalStateException("Not implemented");
-  }
-
-  @Override
   public boolean isOpen() {
     return channelOpen.get();
   }
